+++ internal/metadata/storeiface.go
@@ -116,7 +116,7 @@ type StoreAPI interface {
 	SetBucketStats(bucket string, stat BucketStat) error
 	BackfillBucketStats(bucket string) (BucketStat, error)
 	SetBucketObjectLockEnabled(bucket string, enabled bool) error
-	ListLatestObjectsDelimited(bucket, prefix, delimiter, startAfter string, maxKeys int) ([]ObjectMeta, []string, bool, string, error)
+	ListLatestObjectsDelimited(bucket, prefix, delimiter, startAfter string, maxKeys int) ([]ObjectMeta, []CommonPrefixInfo, bool, string, error)
 	PutPart(uploadID string, part PartInfo) error
 	PutPublicAccessBlock(bucket string, cfg PublicAccessBlockConfig) error
 	PutReplicationConfig(bucket, data string) error